reworked graph plot
The plot draws the lines and vertices now based on the information
in the global data frames. The vertex colormap and glyphmap are linked
and the edges can be colored depending on a column in the edge data frame.

What's still missing are the selection tools for the graph plot and
callbacks. Also controls for the visual appearance of the edges.
Eventually, the arrows should be drawn as arrows if the graph is directed.